feat: Update index.html and main.js for lazy loading of images
The code changes in this commit update the index.html file to include a script tag with the attribute "type" set to "module". This enables the use of ES modules in the JavaScript code. Additionally, the main.js file is modified to implement lazy loading for images, improving the page load performance.

This commit builds upon the previous commits that added web server functionality, updated index.html, and added file verification.
