+++ src/application/resource/index.html
@@ -73,7 +73,7 @@
                 sapien ultricies porta.</p>
         </div>
     </div>
-    <script src="./assets/scripts/main.js"></script>
+    <script type="module" src="./assets/scripts/main.js"></script>
 </body>
 
 </html>
